Inherit from Instance class
