+++ lib/mailchimp_api_v3/interest_category/subclasses.rb
@@ -1,0 +1,14 @@
+require 'mailchimp_api_v3/interest'
+
+module Mailchimp
+  class List
+    class InterestCategory
+      module Subclasses
+        #- VALID_TYPES = %w(checkboxes dropdown radio hidden)
+        def interests(options = {})
+          subclass_from Interests, options
+        end
+      end
+    end
+  end
+end
