--- lib/mailchimp_api_v3/lists.rb
@@ -1,8 +1,0 @@
-require 'mailchimp_api_v3/list'
-
-module Mailchimp
-  class Lists < Collection
-    PATH_KEY = DATA_KEY = 'lists'
-    CHILD_CLASS = List
-  end
-end
